% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/areaPolygonPar.R
\name{areaPolygonPar}
\alias{areaPolygonPar}
\title{Parallel version of areaPolygon from geosphere.}
\usage{
areaPolygonPar(cl, poly)
}
\arguments{
\item{poly}{a \code{SpatialPolygonsDataFrame}}
}
\value{
a vector of areas. Progress is tracked using a log file in the working directory.
}
\description{
Compute the area of a polygon on a sphere. Polygons should not self-intersect. Parallel version of areaPolygon.
}

