% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{build_kml_styles}
\alias{build_kml_styles}
\title{Generate KML style definitions}
\usage{
build_kml_styles(
  hover_labels = FALSE,
  line_color = "ff0000ff",
  line_width = 3,
  point_scale = 0.6,
  label_scale = 0.6
)
}
\arguments{
\item{hover_labels}{Whether labels should only appear on hover}

\item{line_color}{KML color code for lines}

\item{line_width}{Width of lines}

\item{point_scale}{Scale of point icons}

\item{label_scale}{Scale of labels}
}
\value{
KML style string
}
\description{
Generate KML style definitions
}
