% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{calculate_heading}
\alias{calculate_heading}
\title{Calculate bearing between two points}
\usage{
calculate_heading(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1, lon1}{First point coordinates}

\item{lat2, lon2}{Second point coordinates}
}
\value{
Bearing in degrees
}
\description{
Calculate bearing between two points
}
