% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{create_combined_kml}
\alias{create_combined_kml}
\title{Create combined KML file from individual files}
\usage{
create_combined_kml(output_dir)
}
\arguments{
\item{output_dir}{Directory containing individual KML files}
}
\description{
Create combined KML file from individual files
}
