% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{create_line_placemark}
\alias{create_line_placemark}
\title{Generate line placemark KML}
\usage{
create_line_placemark(data, filename)
}
\arguments{
\item{data}{Trackpoint data list}

\item{filename}{GPX filename for labeling}
}
\value{
KML string for line placemark
}
\description{
Generate line placemark KML
}
