% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{create_point_placemarks}
\alias{create_point_placemarks}
\title{Generate point placemarks KML}
\usage{
create_point_placemarks(data, point_frequency = 1)
}
\arguments{
\item{data}{Trackpoint data list}

\item{point_frequency}{Show every Nth point}
}
\value{
KML string for all point placemarks
}
\description{
Generate point placemarks KML
}
