% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{detect_timezone}
\alias{detect_timezone}
\title{Detect timezone from GPS coordinates}
\usage{
detect_timezone(lat, lon)
}
\arguments{
\item{lon}{Vector of longitudes}

\item{decimal_latitude}{Vector of latitudes}
}
\value{
Timezone string
}
\description{
Detect timezone from GPS coordinates
}
