% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{extract_trackpoint_data}
\alias{extract_trackpoint_data}
\title{Extract and process trackpoint data from GPX file}
\usage{
extract_trackpoint_data(gpx_file)
}
\arguments{
\item{gpx_file}{Path to GPX file}
}
\value{
List with lat, lon, time_utc, time_local, local_tz
}
\description{
Extract and process trackpoint data from GPX file
}
