% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{find_gpx_files}
\alias{find_gpx_files}
\title{Find GPX files in a directory}
\usage{
find_gpx_files(folder)
}
\arguments{
\item{folder}{Path to search for GPX files}
}
\value{
Vector of GPX file paths
}
\description{
Find GPX files in a directory
}
