% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{format_time_label}
\alias{format_time_label}
\title{Create formatted time label with timezone}
\usage{
format_time_label(time_local, time_utc, local_tz)
}
\arguments{
\item{time_local}{Local time (POSIXct)}

\item{time_utc}{UTC time (POSIXct)}

\item{local_tz}{Timezone string}
}
\value{
Formatted label string
}
\description{
Create formatted time label with timezone
}
