% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gAreaHoles}
\alias{gAreaHoles}
\title{Area of geometry with holes not correct}
\usage{
gAreaHoles(spgeom, byid = FALSE)
}
\arguments{
\item{spgeom}{sp object as defined in packages sp}

\item{byid}{logical determining if the function should be
applied across subgeometries (TRUE) or the entire object (FALSE)}
}
\value{
Returns the area of the geometry in the units
of the current projection. By definition non-[MULTI]POLYGON
geometries have an area of 0. The area of a POLYGON is
the area of its shell less the area of any holes. Note
that this value may be different from the area slot of
the Polygons class as this value does not subtract
the area of any holes in the geometry.
}
\description{
Fixes holes and calculates the area of the given geometry
}

