% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{geography_coordinate_rounder}
\alias{geography_coordinate_rounder}
\title{Rounds lat lon values based on sensible precision and manage formats
based on accuracy information in an coordinate uncertainty field
in meters}
\usage{
geography_coordinate_rounder(df)
}
\arguments{
\item{df}{a database in SYMBIOTA format}
}
\value{
a dataframe with rounded lat lon values
}
\description{
Rounds lat lon values based on sensible precision and manage formats
based on accuracy information in an coordinate uncertainty field
in meters
}
