% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getGridRowsWeights.R
\name{getGridRowsWeights}
\alias{getGridRowsWeights}
\title{getGridRowsWeights}
\usage{
getGridRowsWeights(geoValue, brickGrid, geoPoly, geoPolyVar, spgridGridSlot)
}
\arguments{
\item{geoValue}{\code{numeric} or \code{character} that
identifies the polygon within geoPoly.}

\item{brickGrid}{The \code{SpatialGridDataFrame} to be overlaid,
converted to a \code{raster brick} object.}

\item{geoPoly}{The \code{SpatialPolygonsDataFrame} containing
the polygon to be overlaid.}

\item{geoPolyVar}{The variable in \code{geoPoly} used to search for
\code{geoValue}.}

\item{spgridGridSlot}{The \code{grid} slot from the given
\code{SpatialGridDataFrame}}
}
\value{
A \code{data.frame} with columns \code{rows} and \code{weights}.
\code{rows} are the rows corresponding to grid cells that the
polygon overlaps with. \code{weights} are the weights for each
of these cells to account for partial overlap.
}
\description{
This utility function gets weights of the rows
of a \code{SpatialGridDataFrame}
(converted to a \code{raster brick} before passing to this
function) corresponding to a particular individual polygon.
}
\details{
This function is used in the process of aggregating when
computation speed is a concern. (the function \code{over()} in
the \code{sp} package, for example, takes a very long time
with no log file
provided). These can then be looped to overlay a \code{SpatialGridDataFrame} on
a \code{SpatialPolygonsDataFrame}. The rows and weights are stored
for repeated overlays of many \code{SpatialGridDataFrame} objects.
rows = c(1,2); weights = c(0.4,0.6) would indicate
that the polygon
}

