% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSmallGrid.R
\name{getSmallGrid}
\alias{getSmallGrid}
\title{Shrinks a rasterbrick for faster computation}
\usage{
getSmallGrid(thisPoly, brick, spgridGridSlot)
}
\arguments{
\item{thisPoly}{polygon that the cells will surround}

\item{brick}{bricked \code{SpatialGridDataFrame} that will be shrunk}

\item{spgridGridSlot}{The \code{grid} slot from the given
\code{SpatialGridDataFrame}}
}
\value{
a raster brick object (subset of brick)
}
\description{
Shrinks a rasterbrick to only those cells that immediately surround
a polygon.
}

