% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools-package.R
\docType{package}
\name{gistools-package}
\alias{gistools}
\alias{gistools-package}
\title{gistools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyingratchet/gistools}
  \item Report bugs at \url{https://github.com/flyingratchet/gistools/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
