% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gridRowsWeightsSpPoly.R
\name{gridRowsWeightsSpPolyPar}
\alias{gridRowsWeightsSpPolyPar}
\title{Extract the rows and weights used for aggregating a SpatialGridDataFrame
to a SpatialPolygonsDataFrame, using \%dopar\% in foreach}
\usage{
gridRowsWeightsSpPolyPar(cl, spgrid, poly, idPoly)
}
\arguments{
\item{cl}{cluster object}

\item{spgrid}{\code{SpatialGridDataFrame} to be aggregated to a polygon}

\item{poly}{\code{SpatialPolygonsDataFrame} that \code{spgrid} is to be
aggregated t.}

\item{idPoly}{vector in \code{poly} that contains a unique polygon
identifier. Pass the column itself to the function - not the name.}
}
\value{
A list of dataframes with columns \code{rows} and \code{weights}.
\code{rows} are the rows corresponding to grid cells that the
polygon overlaps with. \code{weights} are the weights for each
of these cells to account for partial overlap. Each element of the list
corresponds to an individual element of \code{spPoly}.
}
\description{
Computes the rows and weights for overlaying a \code{SpatialGridDataFrame} on
all individual polygons in a \code{SpatialPolygonsDataFrame}. Rows refers
to the rows in the SpatialGridDataFrame; weights are the weights that apply
to the corresponding rows. The procedure is parallelized by default. If
paralellized, progress is reported in a text file named 'log.txt' in the
working directory
}

