% Generated by roxygen2 (4.0.2): do not edit by hand
\name{overGridPoly}
\alias{overGridPoly}
\title{Spatial overlay for grid to polygon}
\usage{
overGridPoly(grid, poly, idPoly = idPoly, gridArcMin, gridVar,
  polyVar = gridVar)
}
\arguments{
\item{grid}{\code{SpatialGridDataFrame} to be overlaid}

\item{poly}{\code{SpatialPolygonsDataFrame} to be added to}

\item{idPoly}{vector in \code{poly} that contains a unique polygon
identifier. Pass the column itself to the function - not the name.}

\item{gridArcMin}{size of the cell in \code{grid}. A future version could remove this.}

\item{gridVar}{\code{character} name of the variable to be extracted from \code{grid}}

\item{polyVar}{\code{character} name of the new variable to be added to \code{poly}. Default is \code{gridVar}.}
}
\value{
a \code{SpatialPolygonsDataFrame}. This will be \code{poly} with an extra column.
}
\description{
Adds a variable to a \code{SpatialPolygonsDataFrame} corresponding to a variable in a SpatialGridDataFrame.
This function is appropriate for a single grid overlay. Use \code{gridRowsWeightsSpPoly} and run a
loop over the grids if there is more than one grid.
}

