% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overGridPoly.R
\name{overGridPolyPar}
\alias{overGridPolyPar}
\title{Spatial overlay for grid to polygon}
\usage{
overGridPolyPar(cl, spgrid, poly, idPoly = idPoly, gridVar,
  polyVar = gridVar)
}
\arguments{
\item{spgrid}{\code{SpatialGridDataFrame} to be overlaid}

\item{poly}{\code{SpatialPolygonsDataFrame} to be added to}

\item{idPoly}{vector in \code{poly} that contains a unique polygon
identifier. Pass the column itself to the function - not the name.}

\item{gridVar}{\code{character} name of the variable to be extracted from \code{spgrid}}

\item{polyVar}{\code{character} name of the new variable to be added to \code{poly}. Default is \code{gridVar}.}

\item{parallel}{cluster object}
}
\value{
a \code{SpatialPolygonsDataFrame}. This will be \code{poly} with an extra column.
}
\description{
Adds a variable to a \code{SpatialPolygonsDataFrame} corresponding to a variable in a SpatialGridDataFrame.
This function is appropriate for a single grid overlay. Use \code{gridRowsWeightsSpPoly} and run a
loop over the grids if there is more than one grid.
}

