% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{process_single_gpx}
\alias{process_single_gpx}
\title{Process a single GPX file into KML}
\usage{
process_single_gpx(gpx_file, output_dir, ...)
}
\arguments{
\item{gpx_file}{Path to GPX file}

\item{output_dir}{Output directory}

\item{...}{Style parameters}
}
\description{
Process a single GPX file into KML
}
