% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistools.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Attach a polygon attribute to a data frame by spatial join}
\usage{
reverse_geocode(df, shapefile, field_name, new_name)
}
\arguments{
\item{df}{A Symbiota-like data.frame with \code{decimal_latitude} and \code{decimal_longitude}}

\item{shapefile}{An sf polygon layer}

\item{field_name}{The single attribute in \code{shapefile} to bring over}

\item{new_name}{Name for the new/updated column on \code{df}}
}
\value{
\code{df} with \code{new_name} column filled where points intersect polygons
}
\description{
Attach a polygon attribute to a data frame by spatial join
}
