% Generated by roxygen2 (4.0.2): do not edit by hand
\name{areaPolygonPar}
\alias{areaPolygonPar}
\title{Parallel version of areaPolygon from geosphere.}
\usage{
areaPolygonPar(poly)
}
\arguments{
\item{poly}{a \code{SpatialPolygonsDataFrame}}
}
\value{
a vector of areas
}
\description{
Compute the area of a polygon on a sphere. Polygons should not self-intersect. Parallel version of areaPolygon.
}

