\name{GISTools-package}
\alias{GISTools-package}
\alias{GISTools}
\title{GISTools}
\description{
In this package, a number of utilities for handling and visualising geographical data of a \dQuote{Spatial} or \dQuote{sf} object - for example choropleth mapping with 'nice' legends.
}

\details{
\tabular{ll}{
Package: \tab GISTools\cr
Type: \tab Package\cr
Version: \tab 1.0-2\cr
Date: \tab 2024-10-03\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}